Задание 6. Обновлены модели и тесты под новую спецификацию
