--- middle-service-app/src/main/java/middle/example/gpb/models/AccountListResponse.java
@@ -1,6 +1,0 @@
-package middle.example.gpb.models;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-public record AccountListResponse(UUID accountId, String accountName, BigDecimal amount) {}
