+++ middle-service-app/src/main/java/middle/example/gpb/models/AccountsListResponseV2.java
@@ -1,0 +1,6 @@
+package middle.example.gpb.models;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public record AccountsListResponseV2(UUID accountId, String accountName, BigDecimal amount) {}
