--- middle-service-app/src/main/java/middle/example/gpb/models/CreateAccountRequest.java
@@ -1,3 +1,0 @@
-package middle.example.gpb.models;
-
-public record CreateAccountRequest(String accountName) {}
