+++ middle-service-app/src/main/java/middle/example/gpb/models/CreateAccountRequestV2.java
@@ -1,0 +1,3 @@
+package middle.example.gpb.models;
+
+public record CreateAccountRequestV2(String accountName) {}
