+++ middle-service-app/src/main/java/middle/example/gpb/models/CreateTransferRequest.java
@@ -2,4 +2,4 @@ package middle.example.gpb.models;
 
 import java.math.BigDecimal;
 
-public record CreateTransferRequest(long fromId, long toId, BigDecimal amount) {}
+public record CreateTransferRequest(String from, String to, BigDecimal amount) {}
