--- middle-service-app/src/main/java/middle/example/gpb/models/CreateUserRequest.java
@@ -1,3 +1,0 @@
-package middle.example.gpb.models;
-
-public record CreateUserRequest(long userId, String userName) {}
