+++ middle-service-app/src/main/java/middle/example/gpb/models/CreateUserRequestV2.java
@@ -1,0 +1,3 @@
+package middle.example.gpb.models;
+
+public record CreateUserRequestV2(long userId, String userName) {}
