--- middle-service-app/src/main/java/middle/example/gpb/models/InnerError.java
@@ -1,5 +1,0 @@
-package middle.example.gpb.models;
-
-import java.util.UUID;
-
-public record InnerError(String message, String type, String code, UUID traceId) {}
