+++ middle-service-app/src/main/java/middle/example/gpb/models/InnerErrorV2.java
@@ -1,0 +1,5 @@
+package middle.example.gpb.models;
+
+import java.util.UUID;
+
+public record InnerErrorV2(String message, String type, String code, UUID traceId) {}
