+++ middle-service-app/src/main/java/middle/example/gpb/models/TransferResponseV2.java
@@ -2,4 +2,4 @@ package middle.example.gpb.models;
 
 import java.util.UUID;
 
-public record TransferResponse(UUID transferId) {}
+public record TransferResponseV2(UUID transferId) {}
