+++ middle-service-app/src/main/java/middle/example/gpb/models/UserResponseV2.java
@@ -2,4 +2,4 @@ package middle.example.gpb.models;
 
 import java.util.UUID;
 
-public record UserResponse(UUID userId) {}
+public record UserResponseV2(UUID userId) {}
